Magic casting in progress

+++ Diablo2/Motor2D/Animation.h
@@ -63,7 +63,7 @@ public:
 
 	int CurrentFrame() const
 	{
-		return current_frame;
+		return (int)current_frame;
 	}
 };
